added register unit update for email same email not update

+++ update.register.unit.php
@@ -79,20 +79,31 @@ if (isset($_POST['form_submit'])) {
     }
     //! Hata yoksa veritabanına kaydet
     if (empty($errors)) {
-        $SORGU = $DB->prepare($sql);
-        $SORGU->bindParam(':form_username', $name);
-        $SORGU->bindParam(':form_email', $email);
-        $SORGU->bindParam(':form_gender', $gender);
-        $SORGU->bindParam(':form_adress', $address);
-        $SORGU->bindParam(':form_phonenumber', $phoneNumber);
-        $SORGU->bindParam(':form_birthdate', $birthDate);
+        //! Kullanıcı e-posta adresini değiştirdiyse, yeni e-posta adresi için veritabanında mevcut bir kullanıcı olup olmadığını kontrol et
+        //? Eğer post edilen email update yapılan kişinin dışında bir kullanıcıda varsa hata mesajı göster
+        $checkEmailQuery = $DB->prepare("SELECT * FROM registerunits WHERE useremail = :email AND userid != :idRegisterUnit");
+        $checkEmailQuery->bindParam(':email', $email);
+        $checkEmailQuery->bindParam(':idRegisterUnit', $id);
+        $checkEmailQuery->execute();
+        $existingUser = $checkEmailQuery->fetch(PDO::FETCH_ASSOC);
+        if ($existingUser) {
+            $errors[] = "This email is already in use.";
+        } else {
+            $SORGU = $DB->prepare($sql);
+            $SORGU->bindParam(':form_username', $name);
+            $SORGU->bindParam(':form_email', $email);
+            $SORGU->bindParam(':form_gender', $gender);
+            $SORGU->bindParam(':form_adress', $address);
+            $SORGU->bindParam(':form_phonenumber', $phoneNumber);
+            $SORGU->bindParam(':form_birthdate', $birthDate);
 
-        $SORGU->bindParam(':idRegisterUnit', $id);
-        $SORGU->execute();
-        echo '<script>';
-        echo 'alert("Register Unit User Update Successful!");';
-        echo 'window.location.href = "update.register.unit.php?idRegisterUnit=' . $registerunits[0]['userid'] . '";';
-        echo '</script>';
+            $SORGU->bindParam(':idRegisterUnit', $id);
+            $SORGU->execute();
+            echo '<script>';
+            echo 'alert("Register Unit User Update Successful!");';
+            echo 'window.location.href = "update.register.unit.php?idRegisterUnit=' . $registerunits[0]['userid'] . '";';
+            echo '</script>';
+        }
     }
 
 }
